Change merge strategy to preserve Plugins when importing Configs
Plugins are merged as `toml.Tree` types. This change adds a Transformer
to the `mergo.merge` function that deals with toml.Tree types.

The merge function overwrites `toml.Tree`s as structs, which has the
side-effect that it gets rid of the Plugin map.

This transformer gets called before the copy, and explicitly preserves
the Plugin values that would otherwise be overwritten.